(fix) Clear emotion state for next session.
(refactor) Move socket event listener initialisation to App.
